refactor(payments): move checkoutSimulator into providers/stripe/ (#36)
checkoutSimulator is Stripe-specific — moves it alongside the other
Stripe implementation files. Updates all import paths accordingly.
