+++ tests/integration/e2e/checkoutSimulator.test.ts
@@ -30,7 +30,7 @@
 import { prisma } from '@/db/client';
 import { getRedisClient } from '@/config/redis';
 import { issueVirtualCard, revealCard } from '@/payments/providers/stripe/cardService';
-import { runSimulatedCheckout } from '@/payments/checkoutSimulator';
+import { runSimulatedCheckout } from '@/payments/providers/stripe/checkoutSimulator';
 import { getStripeClient } from '@/payments/providers/stripe/stripeClient';
 import { IntentStatus } from '@/contracts';
